columns avaliacao

+++ minha_dashboard/views.py
@@ -74,7 +74,7 @@ def home(request):
 
     df=pd.DataFrame(item)
 
-    df['cancelamento_pelo_restaurante'] = df['cancelamento_pelo_restaurante'].astype(float)
+    #df['cancelamento_pelo_restaurante'] = df['cancelamento_pelo_restaurante'].astype(float)
     df['faturamento_real'] = df['faturamento_real'].astype(float)
     df['nps_medio'] = df['nps_medio'].astype(float)
     df['numero_de_comentários'] = df['numero_de_comentários'].astype(float)
